Update ui_fr-fr.json
Not sure if the  '  need to be doubled is json, if it's the case, well, you know how to do it I think ;)
There are some lines that are not done yet, mostly the long ones and the ones that I don't know at all what they're reffering to (Asset for example). I'll let ppl who know what it is do it.